Extract shared code between StreamsAppConfig/ProducerAppConfig to KafkaAppConfig
