+++ kpops/components/base_components/kafka_app.py
@@ -15,6 +15,7 @@ from kpops.component_handlers.helm_wrapper.model import (
 from kpops.component_handlers.helm_wrapper.utils import create_helm_release_name
 from kpops.components.base_components.helm_app import HelmAppValues
 from kpops.components.base_components.models.to_section import (
+    KafkaTopic,
     ToSection,
 )
 from kpops.components.base_components.pipeline_component import PipelineComponent
@@ -52,6 +53,40 @@ class KafkaStreamsConfig(CamelCaseConfigModel, DescConfigModel):
         self, handler: Callable, info: pydantic.SerializationInfo
     ) -> dict[str, Any]:
         return exclude_defaults(self, exclude_by_value(handler(self), None))
+
+
+class KafkaAppConfig(KafkaStreamsConfig):
+    """Kafka Streams settings for all Kafka apps.
+
+    :param extra_output_topics: Extra output topics
+    :param output_topic: Output topic, defaults to None
+    """
+
+    extra_output_topics: dict[str, KafkaTopic] = Field(
+        default={}, description=describe_attr("extra_output_topics", __doc__)
+    )
+    output_topic: KafkaTopic | None = Field(
+        default=None, description=describe_attr("output_topic", __doc__)
+    )
+
+    @pydantic.field_validator("output_topic", mode="before")
+    @classmethod
+    def validate_output_topic(cls, output_topic: Any) -> KafkaTopic | None:
+        if output_topic and isinstance(output_topic, str):
+            return KafkaTopic(name=output_topic)
+        return None
+
+    @pydantic.field_serializer("output_topic")
+    def serialize_topic(self, topic: KafkaTopic | None) -> str | None:
+        if not topic:
+            return None
+        return topic.name
+
+    @pydantic.field_serializer("extra_output_topics")
+    def serialize_extra_output_topics(
+        self, extra_topics: dict[str, KafkaTopic]
+    ) -> dict[str, str]:
+        return {role: topic.name for role, topic in extra_topics.items()}
 
 
 class KafkaAppValues(HelmAppValues):
